+++ packages/pwa-kit-dev/src/utils/script-utils.test.js
@@ -5,16 +5,32 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-import {mkdtemp, rm, writeFile, readJsonSync, readJson} from 'fs-extra'
+import {mkdtemp, rm, writeFile, readJsonSync, readJson, createFile} from 'fs-extra'
+import {execSync} from 'child_process'
 import path from 'path'
 import os from 'os'
 import * as scriptUtils from './script-utils'
+import * as dependencyTreeMockData from './mocks/dependency-tree-mock-data'
 const pkg = readJsonSync(path.join(__dirname, '../../package.json'))
 
+let actualReadJson
+let actualExecSync
+
 jest.mock('fs-extra', () => {
+    const originalModule = jest.requireActual('fs-extra')
+    actualReadJson = originalModule.readJson
     return {
-        ...jest.requireActual('fs-extra'),
+        ...originalModule,
         readJson: jest.fn()
+    }
+})
+
+jest.mock('child_process', () => {
+    const originalModule = jest.requireActual('child_process')
+    actualExecSync = originalModule.execSync
+    return {
+        ...originalModule,
+        execSync: jest.fn()
     }
 })
 
@@ -25,6 +41,10 @@ describe('scriptUtils', () => {
     beforeEach(async () => {
         process.env = {...originalEnv}
         tmpDir = await mkdtemp(path.join(os.tmpdir(), 'scriptUtils-tests'))
+        // This is a workaround for jest.spyOn(), since I guess it doesn't work with our imports?
+        // In any case, using the actual implementation by default prevents subtle bugs in tests.
+        readJson.mockReset().mockImplementation(actualReadJson)
+        execSync.mockReset().mockImplementation(actualExecSync)
     })
 
     afterEach(async () => {
@@ -134,7 +154,99 @@ describe('scriptUtils', () => {
         })
     })
 
-    jest.unmock('fs-extra')
+    describe('getLowestPackageVersion', () => {
+        test('should work when major version is different', async () => {
+            const lowestVersion = await scriptUtils.getLowestPackageVersion(
+                '@salesforce/pwa-kit-react-sdk',
+                dependencyTreeMockData.differentMajorVersions
+            )
+            expect(lowestVersion).toBe('9.0.0')
+        })
+
+        test('should work when minor version is different', async () => {
+            const lowestVersion = await scriptUtils.getLowestPackageVersion(
+                '@salesforce/pwa-kit-react-sdk',
+                dependencyTreeMockData.differentMinorVersions
+            )
+            expect(lowestVersion).toBe('1.9.0')
+        })
+
+        test('should work when patch version is different', async () => {
+            const lowestVersion = await scriptUtils.getLowestPackageVersion(
+                '@salesforce/pwa-kit-react-sdk',
+                dependencyTreeMockData.differentPatchVersions
+            )
+            expect(lowestVersion).toBe('1.0.9')
+        })
+
+        test('should work when version contains pre-release version', async () => {
+            const lowestVersion = await scriptUtils.getLowestPackageVersion(
+                '@salesforce/pwa-kit-react-sdk',
+                dependencyTreeMockData.preReleaseVersion
+            )
+            expect(lowestVersion).toBe('1.0.0-beta')
+        })
+
+        test('should work when package is deduped', async () => {
+            const lowestVersion = await scriptUtils.getLowestPackageVersion(
+                '@salesforce/pwa-kit-react-sdk',
+                dependencyTreeMockData.dedupedVersion
+            )
+            expect(lowestVersion).toBe('1.0.0')
+        })
+
+        test("should return 'unknown' when package not found", async () => {
+            const lowestVersion = await scriptUtils.getLowestPackageVersion(
+                '@salesforce/pwa-kit-react-sdk',
+                dependencyTreeMockData.noPwaKitPackages
+            )
+            expect(lowestVersion).toBe('unknown')
+        })
+    })
+
+    describe('getPwaKitDependencies', () => {
+        test('should return pwa-kit packages with unknown version if not in dependency tree', async () => {
+            const dependencies = await scriptUtils.getPwaKitDependencies(
+                dependencyTreeMockData.noPwaKitPackages
+            )
+            expect(Object.keys(dependencies)).toHaveLength(3)
+            expect(dependencies).toHaveProperty('@salesforce/pwa-kit-react-sdk', 'unknown')
+            expect(dependencies).toHaveProperty('@salesforce/pwa-kit-runtime', 'unknown')
+            expect(dependencies).toHaveProperty('@salesforce/pwa-kit-dev', 'unknown')
+        })
+
+        test('should return pwa-kit packages with version in dependency tree', async () => {
+            const dependencies = await scriptUtils.getPwaKitDependencies(
+                dependencyTreeMockData.includesPwaKitPackages
+            )
+            expect(Object.keys(dependencies)).toHaveLength(3)
+            expect(dependencies).toHaveProperty('@salesforce/pwa-kit-react-sdk', '1.0.0')
+            expect(dependencies).toHaveProperty('@salesforce/pwa-kit-runtime', '1.0.0')
+            expect(dependencies).toHaveProperty('@salesforce/pwa-kit-dev', '1.0.0')
+        })
+    })
+
+    describe('getProjectDependencyTree', () => {
+        let originalCwd
+        beforeAll(() => {
+            originalCwd = process.cwd()
+        })
+        afterEach(() => process.chdir(originalCwd))
+        test('works in retail-react-app', async () => {
+            expect(await scriptUtils.getProjectDependencyTree()).toMatchObject({
+                name: '@salesforce/pwa-kit-dev',
+                version: pkg.version,
+                dependencies: expect.any(Object)
+            })
+        }, 10_000) // This test can take a while on CI
+        test('returns nothing if an error occurs', async () => {
+            execSync.mockImplementation(() => {
+                throw new Error('npm ls did not work')
+            })
+            expect(await scriptUtils.getProjectDependencyTree()).toBeNull()
+        })
+    })
+
     describe('defaultMessage', () => {
         test('works', async () => {
             const mockGit = {branch: () => 'branch', short: () => 'short'}
@@ -222,6 +334,7 @@ describe('scriptUtils', () => {
 
         test('should archive a bundle', async () => {
             readJson.mockReturnValue(pkg)
+            execSync.mockReturnValue(JSON.stringify(dependencyTreeMockData.noPwaKitPackages))
             const message = 'message'
             const bundle = await scriptUtils.createBundle({
                 message,
@@ -237,6 +350,8 @@ describe('scriptUtils', () => {
             expect(bundle.ssr_parameters).toEqual({})
             expect(bundle.ssr_only).toEqual(['ssr.js'])
             expect(bundle.ssr_shared).toEqual(['ssr.js', 'static/favicon.ico'])
+            expect(bundle.bundle_metadata).toHaveProperty('dependencies')
+            expect(bundle.bundle_metadata).toHaveProperty('cc_overrides')
 
             // De-code and re-encode gives the same result, to show that it *is* b64 encoded
             expect(Buffer.from(bundle.data, 'base64').toString('base64')).toEqual(bundle.data)
@@ -274,6 +389,7 @@ describe('scriptUtils', () => {
             'should push a built bundle and handle status codes (%p)',
             async ({projectSlug, targetSlug, expectedURL, status}) => {
                 readJson.mockReturnValue(pkg)
+                execSync.mockReturnValue(JSON.stringify(dependencyTreeMockData.noPwaKitPackages))
                 const message = 'message'
                 const bundle = await scriptUtils.createBundle({
                     message,
@@ -401,6 +517,21 @@ describe('scriptUtils', () => {
                 message: '\tThis is a test log message 550e8400',
                 shortRequestId: '550e8400'
             })
+        })
+    })
+
+    describe('walkDir', () => {
+        const files = ['a', 'b/1', 'b/2', 'c/d/e'].map(path.normalize)
+        beforeEach(async () => {
+            await Promise.all(files.map(async (file) => await createFile(path.join(tmpDir, file))))
+        })
+        test('finds all files in a directory', async () => {
+            const result = await scriptUtils.walkDir(tmpDir, tmpDir)
+            expect([...result]).toEqual(files)
+        })
+        test('returns file relative to specified path', async () => {
+            const result = await scriptUtils.walkDir(tmpDir, '/')
+            expect([...result]).toEqual(files.map((f) => path.join(tmpDir, f)))
         })
     })
 })